requirements.txt
absl-py==2.1.0
attrs==23.2.0
cffi==1.16.0
comtypes==1.3.1
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.3.1
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mediapipe==0.10.11
ml-dtypes==0.3.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.2.0
protobuf==3.20.3
psutil==5.9.8
pyalsaaudio==0.10.0
PyAutoGUI==0.9.54
pycaw==20240210
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
scipy==1.12.0
semantic-version==2.10.0
setuptools-rust==1.9.0
six==1.16.0
sounddevice==0.4.6
sympy==1.12
torch==2.2.1
triton==2.2.0
typing_extensions==4.10.0
